Overview: Implenting a Base Schema for Posts and Creating Controllers to make Sucessful Requst to Database
Now if the user has a URI to the databse, should be able to upload posts. I created a post schema, which is kinda a dummy right now so I can test controlleres and the db. Still need to decide on all aspects of the document, something that may require communciating with frontend on what info is requried. In the controllers folder, postController should contain functions that the router can call for get/post requests. Some of the request in the router are still dummy or not neccesary, this commit mainly ensures a base template is not working for routes and requests to the db. To test out the routes, I used Postman, an app that allows you to make requests to the server.If you wanna try it out maybe google or chatgpt how to use postman, shouldn't be to hard to figure out. Also additioanl note, the request to db should currently I belive in Json format? Still kinda new to this so not completely sure.